Fixup comparison of integrers of different signs warning

+++ core/src/impl/Kokkos_ConcurrentBitset.hpp
@@ -144,7 +144,8 @@ struct concurrent_bitset {
       // Failed race to set the selected bit
       // Find a new bit to try.
 
-      const int j = (prev == -1 ? -1 : Kokkos::countr_one(prev));
+      const int j =
+          (prev == static_cast<uint32_t>(-1) ? -1 : Kokkos::countr_one(prev));
 
       if (0 <= j) {
         bit = (word << bits_per_int_lg2) | uint32_t(j);
@@ -233,7 +234,8 @@ struct concurrent_bitset {
       // Failed race to set the selected bit
       // Find a new bit to try.
 
-      const int j = (prev == -1 ? -1 : Kokkos::countr_one(prev));
+      const int j =
+          (prev == static_cast<uint32_t>(-1) ? -1 : Kokkos::countr_one(prev));
 
       if (0 <= j) {
         bit = (word << bits_per_int_lg2) | uint32_t(j);
